feat: Move DateInput stories back into DatePicker
